# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: FlowAccess API specification
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /extension/token:
    get:
      operationId: getExtensionToken
      tags: [extension]
      summary: Get or create API token for extension
      description: Returns or creates an API token for the current user (used by Chrome extension)
      responses:
        "200":
          description: API token
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiTokenResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /extension/me:
    get:
      operationId: getExtensionUser
      tags: [extension]
      summary: Get user info via API token
      description: Verifies an API token (X-API-Token header) and returns user info
      parameters:
        - in: header
          name: X-API-Token
          required: true
          schema:
            type: string
          description: API token issued via GET /extension/token
      responses:
        "200":
          description: User info
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserProfile"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /extension/inject:
    post:
      operationId: injectSession
      tags: [extension]
      summary: Inject a Google session
      description: Verifies API token, picks least-recently-used active session, deducts 1 credit, logs usage, returns cookie data
      parameters:
        - in: header
          name: X-API-Token
          required: true
          schema:
            type: string
          description: API token issued via GET /extension/token
      responses:
        "200":
          description: Cookie data for injection
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/InjectResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: No credits remaining
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: No active session available
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required:
        - error
...
    UserProfile:
      type: object
      properties:
        id:
          type: integer
        clerkUserId:
          type: string
        email:
          type: string
        isAdmin:
          type: boolean
        planId:
          type: ["integer", "null"]
        planName:
          type: ["string", "null"]
        creditsTotal:
          type: integer
        creditsUsed:
          type: integer
        creditsRemaining:
          type: integer
        subscriptionStartedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - clerkUserId
        - email
        - isAdmin
        - creditsTotal
        - creditsUsed
        - creditsRemaining
        - createdAt
    ApiTokenResponse:
      type: object
      properties:
        token:
          type: string
      required:
        - token
    InjectResponse:
      type: object
      properties:
        cookieData:
          type: string
        creditsRemaining:
          type: integer
      required:
        - cookieData
        - creditsRemaining
...